move utils to crypto, add utils with other common functions
