+++ package.json
@@ -29,7 +29,10 @@
     "webpack-cli": "^3.1.2"
   },
   "dependencies": {
-    "http-server": "^0.11.1"
+    "bluebird": "^3.5.3",
+    "http-server": "^0.11.1",
+    "hyperdb": "^3.5.0",
+    "random-access-idb": "^1.2.0"
   },
   "repository": {
     "type": "git",
