+++ src/index.js
@@ -1,4 +1,6 @@
 import utils from './utils'
+import crypto from './crypto'
 import errors from './errors'
 module.exports.utils = utils
 module.exports.errors = errors
+module.exports.crypto = crypto
