+++ test/index.html
@@ -20,7 +20,7 @@
     mocha.setup('bdd')
     mocha.checkLeaks();
   </script>
-  <script src="./utils.js"></script>
+  <script src="./crypto.js"></script>
   <script>
     mocha.run();
   </script>
